requirements.txt
discord-py-interactions
datetime
python-dotenv
pynacl
pydub
requests